Update covers script as auto mode was deprecated
Auto mode is deprecated as browsers are moving towards a default strict-origin-when-cross-origin referrer policy, which makes it impossible to automatically derive the actual user and list type. If inference fails, it will try to fall back to serving all possible entries. If you use the 'more' preset this will also fail - you will have to change to use one of the other options listed above; either specify your username in full, or use one of the specific 'all' routes for either anime or manga.